Fix a lot of warnings (for fussy compilers) and one always-true check

+++ chocolate-doom/src/net_server.c
@@ -869,7 +869,7 @@ static void NET_SV_ParseGameData(net_packet_t *packet, net_client_t *client)
     unsigned int seq;
     unsigned int num_tics;
     unsigned int nowtime;
-    int i;
+    size_t i;
     int player;
     int resend_start, resend_end;
     int index;
@@ -1020,7 +1020,7 @@ static void NET_SV_ParseResendRequest(net_packet_t *packet, net_client_t *client
 {
     unsigned int start, last;
     unsigned int num_tics;
-    int i;
+    unsigned int i;
 
     // Read the starting tic and number of tics
